Update environment initialization
Since X sources profile and TTYs do not if fish is a login shell and
TTYs on some of my systems do not source ~/.pam_environment I have to
use .profile for a general interface for seeting environment variables.
Because of inconsistencies between systems, this is likely to be changed
or tweaked in the future.